pyproject.toml
[project]
name = "video-notes-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "htmldocx>=0.0.6",
    "pandas>=2.3.3",
    "python-docx>=1.2.0",
    "st-quill-dark-mode>=0.0.3",
    "streamlit>=1.52.1",
    "streamlit-player>=0.1.5",
    "streamlit-quill>=0.0.3",
]
